[api] bug Fix to Locations Symptoms
