+++ Hack/Backend/routes/LocationUserRoutes.js
@@ -7,8 +7,7 @@ var location_user_controller = require("../controllers/LocationUserController.js
 
 // router.get("/api/user_locations",  location_user_controller.get_all_location_users);
 router.get("/api/user_locations/:id", verifyToken.verifyToken, location_user_controller.get_location_user_by_id);
-router.get("/api/user_locations/location/:location_id", verifyToken.verifyToken, location_user_controller.get_by_location_id);
-
+router.post("/api/locations_symptoms", verifyToken.verifyToken, location_user_controller.get_all_locations_with_symptoms);
 router.get("/api/user_locations/user/:user_id", verifyToken.verifyToken, location_user_controller.get_by_user_id);
 
 router.post("/api/user_locations", verifyToken.verifyToken, location_user_controller.post_location_user);
